💥 Started partitioning classes to move to SQLLite - SQL Lite Nuget package from Microsoft is added
