+++ UPBot Code/Commands/Stats.cs
@@ -172,10 +172,4 @@ public class Stats : BaseCommandModule {
     }
   }
 
-  public class Mentioner {
-    public ulong member;
-    public int num;
-  }
-
-
 }
