+++ UPBot Code/DataClasses/LastRequestByMember.cs
@@ -1,0 +1,7 @@
+﻿using System;
+
+public class LastRequestByMember {
+  public ulong MemberId { get; set; }
+  public DateTime DateTime { get; set; }
+  public int Num { get; set; }
+}
