+++ UPBot Code/DataClasses/Mentioner.cs
@@ -1,0 +1,4 @@
+﻿public class Mentioner {
+  public ulong member;
+  public int num;
+}
