+++ UPBot Code/DataClasses/Reputation.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+public class Reputation { // 16 bytes
+  public ulong user;              // 8
+  public ushort reputation;       // 2
+  public ushort fun;              // 2
+  public DateTime startTracking;  // 4
+}
